+++ OndselSolver/FullVector.h
@@ -20,8 +20,8 @@ namespace MbD {
 	public:
 		FullVector() : Array<T>() {}
 		FullVector(std::vector<T> vec) : Array<T>(vec) {}
-		FullVector(int count) : Array<T>(count) {}
-		FullVector(int count, const T& value) : Array<T>(count, value) {}
+		FullVector(size_t count) : Array<T>(count) {}
+		FullVector(size_t count, const T& value) : Array<T>(count, value) {}
 		FullVector(typename std::vector<T>::iterator begin, typename std::vector<T>::iterator end) : Array<T>(begin, end) {}
 		FullVector(std::initializer_list<T> list) : Array<T>{ list } {}
 		double dot(std::shared_ptr<FullVector<T>> vec);
